backend/requirements.txt
langflow
langchain
instaloader
fastapi
google-generativeai